+++ example/bounce/view/board.dart
@@ -3,10 +3,10 @@ part of bounce;
 class Board extends Object with Surface {
   Board(CanvasElement canvas) {
     this.canvas = canvas;
-    movablePieces = new MovablePieces();
-    movablePieces.create(32);
-    movablePieces.randomInit();
-    movablePieces.forEach((MovablePiece p) {
+    pieces = new Pieces();
+    pieces.create(32);
+    pieces.randomInit();
+    pieces.forEach((Piece p) {
       p.shape = PieceShape.CIRCLE;
       p.space = new Area.from(width, height);
     });