+++ example/breakout/model/gear.dart
@@ -1,6 +1,6 @@
 part of breakout;
 
-class Ball extends MovablePiece {
+class Ball extends Object with Piece {
   static const num defaultRadius = 10;
 
   num _radius;
@@ -20,7 +20,7 @@ class Ball extends MovablePiece {
   }
 }
 
-class Racket extends MovablePiece {
+class Racket extends Object with Piece {
   static const num defaultWidth = 75;
   static const num defaultHeight = 10;