+++ example/chaos/view/board.dart
@@ -4,11 +4,11 @@ class Board extends Object with Surface {
   Board(CanvasElement canvas) {
     this.canvas = canvas;
     avoidCollisions = true;
-    movablePieces = new MovablePieces();
-    movablePieces.create(16);
-    movablePieces.randomExtraInit();
+    pieces = new Pieces();
+    pieces.create(16);
+    pieces.randomExtraInit();
     canvas.onMouseDown.listen((MouseEvent e) {
-      movablePieces.onOff();
+      pieces.onOff();
     });
   }
 }