+++ example/move/game.dart
@@ -7,8 +7,8 @@ import 'package:boarding/boarding.dart';
 main() {
   var surface = new Surface();
   surface.canvas = querySelector('#canvas');
-  surface.movablePieces = new MovablePieces();
-  surface.movablePieces.create(8);
-  surface.movablePieces.randomInit();
+  surface.pieces = new Pieces();
+  surface.pieces.create(8);
+  surface.pieces.randomInit();
   surface.draw();
 }